MC-48968 change price value to double

+++ src/main/java/ru/moysklad/remap_1_2/entities/Price.java
@@ -26,5 +26,5 @@ public class Price {
     /**
      * Значение (в копейках)
      */
-    private Long value;
+    private Double value;
 }
